+++ src/app.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
 
+import { HealthModule } from './health/health.module';
+import { InfoModule } from './info/info.module';
 import { OperationsModule } from './operations/operations.module';
 
 @Module({
-  imports: [OperationsModule],
+  imports: [OperationsModule, HealthModule, InfoModule],
 })
 export class AppModule {}